monster.rb, brave.rbファイル作成
